+++ CHANGELOG.rst
@@ -4,7 +4,12 @@ CHANGELOG
 4.7.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Provide `request_indexer` setting to be able to override how we handle
+  indexing data
+  [vangheem]
+
+- Provide `connection_settings` on the request object with `tests.utils.get_container`
+  [vangheem]
 
 
 4.7.1 (2019-04-26)